requirements_peak.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
arfpy==0.1.1
async-timeout==5.0.1
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.5.0
be_great==0.0.9
blis==1.2.1
boto3==1.39.15
botocore==1.39.15
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.2
copulas==0.12.3
ctgan==0.11.0
cycler==0.12.1
cymem==2.0.11
dacite==1.9.2
datasets==4.0.0
decaf-synthetic-data==0.1.7
deepecho==0.7.0
dill==0.3.8
exceptiongroup==1.3.0
Faker==37.4.2
fastai==2.7.19
fastcore==1.7.29
fastdownload==0.0.7
fastprogress==1.0.3
feather-format==0.4.1
fflows==0.0.3
filelock==3.18.0
fonttools==4.59.0
formulaic==1.2.0
frozenlist==1.7.0
fsspec==2025.3.0
geomloss==0.2.6
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
gretel-client==0.25.2
grpcio==1.74.0
grpcio-status==1.71.2
h5py==3.14.0
htmlmin==0.1.12
httplib2==0.22.0
huggingface-hub==0.34.2
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
importlib-metadata==8.7.0
inflate64==1.0.3
inflection==0.5.1
iniconfig==2.1.0
interface-meta==1.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
keopscore==2.3
kiwisolver==1.4.8
langcodes==3.5.0
language-data==1.3.0
lifelines==0.29.0
lightning-utilities==0.15.0
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
monai==1.4.0
mpmath==1.3.0
multidict==6.6.3
multimethod==1.12
multiprocess==0.70.16
multivolumefile==0.2.3
murmurhash==1.0.13
narwhals==2.0.0
networkx==2.8.8
nflows==0.14
numba==0.61.0
numpy==1.26.4
opacus==1.5.4
opt-einsum==3.4.0
optuna==4.4.0
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pgmpy==0.1.26
phik==0.12.5
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
puremagic==1.30
py7zr==1.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybcj==1.0.6
pybind11==3.0.0
pycox==0.3.0
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
Pygments==2.19.2
pykeops==2.3
pyparsing==3.2.3
pyppmd==1.2.0
pytest==8.4.1
pytest-mock-3.14.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pyts==0.13.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzstd==0.17.0
rdt==1.17.1
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.1.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sdmetrics==0.22.0
sdv==1.24.1
seaborn==0.13.2
shap==0.48.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.3.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
statsmodels==0.14.5
sympy==1.14.0
synthcity==0.2.12
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
texttable==1.7.0
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
torch==2.2.2
torchmetrics==1.8.0
torchtext==0.17.2
torchtuples==0.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.54.0
tsai==0.4.0
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
visions==0.8.1
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
win32-setctime==1.2.0
wordcloud==1.9.4
wrapt==1.17.2
xgboost==2.1.4
xgbse==0.3.3
xxhash==3.5.0
yarl==1.20.1
ydata-profiling==4.16.1
zipp==3.23.0